tweaks from MEW to get VTBuilder to build/run on Linux / wxGTK
git-svn-id: http://vtp.googlecode.com/svn/trunk@883 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainApps/VTBuilder/ResampleDlg.cpp
@@ -123,8 +123,8 @@ void ResampleDlg::OnSpacingXY( wxCommandEvent &event )
         return;
 
     TransferDataFromWindow();
-    m_iSizeX = (m_fAreaX / m_fSpacingX)+1;
-    m_iSizeY = (m_fAreaY / m_fSpacingY)+1;
+    m_iSizeX = (int) (m_fAreaX / m_fSpacingX)+1;
+    m_iSizeY = (int) (m_fAreaY / m_fSpacingY)+1;
 
     m_bSetting = true;
     TransferDataToWindow();
